Criando coluna status, e corrigindo rederização de updtade e insert item adicional
